refactor/page specific components to page folder
