+++ frontend/src/components/FormHelpers/ColorPickerDialog.vue
