+++ frontend/src/pages/Admin/Settings/CreatePageDialog.vue
