+++ frontend/src/pages/Admin/Settings/HomePageSettings.vue
