+++ frontend/src/pages/Admin/Theme/NewThemeDialog.vue
